+++ src/common/decimal_type.cpp
@@ -10,11 +10,12 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "common/decimal_value.h"
-#include "common/boolean_value.h"
-#include "common/varlen_value.h"
+#include "common/decimal_type.h"
+
 #include <iostream>
 #include <cmath>
+#include "common/boolean_type.h"
+#include "common/varlen_type.h"
 
 namespace peloton {
 namespace common {
