+++ src/common/timestamp_type.cpp
@@ -10,9 +10,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "common/timestamp_value.h"
-#include "common/boolean_value.h"
-#include "common/varlen_value.h"
+#include "common/timestamp_type.h"
+
+#include "common/boolean_type.h"
+#include "common/varlen_type.h"
 
 namespace peloton {
 namespace common {
