+++ src/common/type.cpp
@@ -11,13 +11,14 @@
 //===----------------------------------------------------------------------===//
 
 #include "common/type.h"
+
+#include "common/array_type.h"
+#include "common/boolean_type.h"
+#include "common/decimal_type.h"
+#include "common/numeric_type.h"
+#include "common/timestamp_type.h"
+#include "common/varlen_type.h"
 #include "common/value.h"
-#include "common/array_value.h"
-#include "common/numeric_value.h"
-#include "common/decimal_value.h"
-#include "common/boolean_value.h"
-#include "common/varlen_value.h"
-#include "common/timestamp_value.h"
 #include "common/exception.h"
 #include "common/varlen_pool.h"
