+++ src/common/value.cpp
@@ -11,11 +11,12 @@
 //===----------------------------------------------------------------------===//
 
 #include "common/value.h"
-#include "common/boolean_value.h"
-#include "common/decimal_value.h"
-#include "common/numeric_value.h"
-#include "common/timestamp_value.h"
-#include "common/varlen_value.h"
+
+#include "common/boolean_type.h"
+#include "common/decimal_type.h"
+#include "common/numeric_type.h"
+#include "common/timestamp_type.h"
+#include "common/varlen_type.h"
 
 namespace peloton {
 namespace common {
