+++ src/include/common/array_type.h
