+++ src/include/common/numeric_type.h
