+++ src/include/common/value_factory.h
@@ -1,14 +1,15 @@
 #pragma once
 
 #include <stdexcept>
-#include "common/array_value.h"
-#include "common/boolean_value.h"
-#include "common/decimal_value.h"
+
+#include "common/array_type.h"
+#include "common/boolean_type.h"
 #include "common/macros.h"
-#include "common/numeric_value.h"
-#include "common/timestamp_value.h"
-#include "common/varlen_value.h"
 #include "common/serializer.h"
+#include "common/decimal_type.h"
+#include "common/numeric_type.h"
+#include "common/timestamp_type.h"
+#include "common/varlen_type.h"
 
 namespace peloton {
 namespace common {
