+++ test/common/array_value_test.cpp
@@ -15,11 +15,12 @@
 #include <iostream>
 #include <cstdint>
 #include <cmath>
-#include "common/numeric_value.h"
-#include "common/boolean_value.h"
-#include "common/decimal_value.h"
-#include "common/varlen_value.h"
-#include "common/array_value.h"
+
+#include "../../src/include/common/array_type.h"
+#include "../../src/include/common/boolean_type.h"
+#include "../../src/include/common/decimal_type.h"
+#include "../../src/include/common/numeric_type.h"
+#include "../../src/include/common/varlen_type.h"
 #include "common/harness.h"
 
 namespace peloton {
